+++ packages/react/shared/src/components/data/WebtrekkAdvancedData.js
@@ -31,7 +31,10 @@ WebtrekkAdvancedData.propTypes = {
     }),
     userIdentification: PropTypes.shape({
         enableOptOut: PropTypes.bool,
+        enableAnonymousFunction: PropTypes.bool,
+        anonymousOptIn: PropTypes.bool,
         optOutCookieName: PropTypes.string,
+        anonymousCookieName: PropTypes.string,
         suppressParameter: PropTypes.arrayOf(PropTypes.string)
     })
 };
@@ -60,7 +63,10 @@ WebtrekkAdvancedData.defaultProps = {
     },
     userIdentification: {
         enableOptOut: null,
+        enableAnonymousFunction: null,
+        anonymousOptIn: null,
         optOutCookieName: null,
+        anonymousCookieName: null,
         suppressParameter: null
     }
 };